Format strings that work both on 32-bit and 64-bit platforms.
[FIXED]     When formatting integers and longs, make it work on both
            32-bit and 64-bit platforms.

+++ Pearl/PearlConfig.m
@@ -329,7 +329,7 @@
 
 - (long)gameRandom:(NSUInteger)scope {
 
-    NSAssert2(scope < PearlMaxGameRandom, @"Scope (%u) must be < %u", scope, PearlMaxGameRandom);
+    NSAssert2(scope < PearlMaxGameRandom, @"Scope (%lu) must be < %u", (long)scope, PearlMaxGameRandom);
 
     @synchronized (self) {
         srandom(_gameRandomSeeds[scope]++);
